ECS: Entity component system : all component, not complete [+]position.(hpp/cpp) : changing type of _y and _x to unsigned short; [+]bonus.hpp : files for ECS bonus (no comment in files + empty ECS); [+]controlable(.hpp/.cpp) : files for ECS controlable (no comment in files); [+]hitbox(.hpp/.cpp) : files for ECS hitbox (no comment in files); [+]pattern.hpp : files for ECS pattern (no comment in files + empty ECS); [+]rotation(.hpp/.cpp) : files for ECS rotation (no comment in files); [+]size(.hpp/.cpp) : files for ECS size (no comment in files); [+]sound.hpp : files for ECS sound (no comment in files + empty ECS); [+]speed(.hpp/.cpp) : files for ECS speed (no comment in files); [+]text(.hpp/.cpp) : files for ECS text (no comment in files); [+]texture.hpp : files for ECS texture (no comment in files + empty ECS); [+]weapon.hpp : files for ECS weapon (no comment in files + empty ECS);

+++ shared/ECS/include/position.hpp
@@ -1,11 +1,10 @@
 #include "IComp.hpp"
-#include <utility>
 
 namespace ECS {
 
 class Position : public IComp {
     public:
-        explicit Position(int x, int y) noexcept;
+        explicit Position(unsigned short x, unsigned short y) noexcept;
         Position(const Position& pos) noexcept;
         explicit Position(Position&&) noexcept = delete;
         ~Position() noexcept override = default;
@@ -13,13 +12,13 @@ class Position : public IComp {
         Position &operator=(const Position&) const noexcept = delete;
         Position &operator=(Position&&) const noexcept = delete;
 
-        [[nodiscard]] int getPosition_x() const noexcept;
-        [[nodiscard]] int getPosition_y() const noexcept;
-        void setPosition_x(int x) noexcept;
-        void setPosition_y(int y) noexcept;
+        [[nodiscard]] unsigned short getPosition_x() const noexcept;
+        [[nodiscard]] unsigned short getPosition_y() const noexcept;
+        void setPosition_x(unsigned short x) noexcept;
+        void setPosition_y(unsigned short y) noexcept;
 
     private:
-        int _x;
-        int _y;
+        unsigned short _x;
+        unsigned short _y;
 };
 };
